Fix layout of main module MP.W.
Apply new token scanner for the Index.  Thanks to Martin Ruckert
(@ruckertm) for the preliminary version of these macros.  They already
work for MP.W, because MP.CH is applied with TIE, not with CWEAVE.
Index entries marked as 'changed' will be supported soon.

Thanks again to Martin for the new version 'newidx5.tex', now updated in
the MP.CH changefile.